servo: Merge #5935 - Upgrade Rust (from servo:rustup_2015-04-25); r=Ms2ger
r? everybody

Source-Repo: https://github.com/servo/servo
Source-Revision: 49aed6555dbc008c1a378c5cbb303f5467232b6b

UltraBlame original commit: 766c983c2579b5ed4989f396cfe2c372a8e9a859

+++ servo/components/script/layout_interface.rs
@@ -16,7 +16,6 @@ use profile_traits::mem::{Reporter, ReportsChan};
 use script_traits::{ScriptControlChan, OpaqueScriptLayoutChannel, UntrustedNodeAddress};
 use std::any::Any;
 use std::sync::mpsc::{channel, Receiver, Sender};
-use std::boxed::BoxAny;
 use style::animation::PropertyAnimation;
 use style::media_queries::MediaQueryList;
 use style::stylesheets::Stylesheet;
@@ -88,7 +87,7 @@ pub struct HitTestResponse(pub UntrustedNodeAddress);
 pub struct MouseOverResponse(pub Vec<UntrustedNodeAddress>);
 
 
-#[derive(PartialEq, Copy, Debug)]
+#[derive(PartialEq, Copy, Clone, Debug)]
 pub enum ReflowGoal {
     
     ForDisplay,
@@ -185,15 +184,15 @@ pub struct Animation {
     
     pub property_animation: PropertyAnimation,
     
-    pub start_time: f64,
+    pub start_time: f32,
     
-    pub end_time: f64,
+    pub end_time: f32,
 }
 
 impl Animation {
     
     #[inline]
-    pub fn duration(&self) -> f64 {
+    pub fn duration(&self) -> f32 {
         self.end_time - self.start_time
     }
 }
